requirements.txt
numpy>=1.21.0
openai>=1.0.0
python-dotenv>=0.19.0
